Redirected starter screen to login screen
